DHCP client that does full lease acquisition.
Previous proof-of-concept only sent a single DISCOVER. This enables
the client to receive and handle OFFER and ACK response from the
server and send REQUEST after OFFER.

Also, add second command line argument for the address of the server.
If not provided, it will use the broadcast address of the network
interface it binds to. If the server address is specified and it is
a broadcast address, it will have the same behavior. If the specified
address is not broadcast then the messages from the client will use
that address to talk to the server rather than sending to broadcast.