Switch from lookup/args to extra context column
Instead of using extra columns for header and tlv defintions, pass extra
context in a context parameter. This includes the readers/writers, TLV
type and length sizing, bitfield specs, etc.

Remove the support for function definition in the default and length
columns. It's neat, but it moves away from more declarative definition
which is the real goal anyways.