Add ping. Addrs format change. Fix header writing.
Add ping example. This uses the raw-socket module. Add ping example to
docker-compose test. Change the addrs parsing to return the more natural
format for values when reading. Currently ipv4 and mac are just
basically raw readers. By returning the text formats it greatly
simplifies things for the DHCP and and ping examples which now do less
conversion back and forth. Add support for :msg-map in the context
parameter so that reader functions can refer to earlier read fields.

Split write-headers headers into a write-headers* alternate version
which returns the written ending offset rather than the sliced buffer as
write-headers does. This allows it to be used as a writer function.