Add :loookup type. Improve/simplify pcap/dhcp/icmp.
The protocols implementations all had a common pattern of translating
the top-level message type using a lookup to that had keys for both
direction lookup. Add a :lookup type that takes does this based on the
value of the current field. The extra-context contains :lookup-type
which specifies the type of the field being used to do the lookup and
the :lookup map which has key/values in both directions. Add
a fields/list->lookup which simplifies building the lookup maps.

Update the dhcp, icmp, and pcap protocol implementations to use :lookup.
This greatly simplifies some of the definitions. It also remove the need
to '(set! *warn-on-infer* false)' in all but the fields.cljs file
because that was the last use of direct scalar reader calls.