[json_pool] EDN cfg file with field/option override.
If the optional second argument specifies an EDN config file then this
is loaded and merged with the settings. It can override any of the
settings apart from the callbacks and the pool atom. In addition, if
a :fields key is specified then any fields and options under that key
override fields and options of DHCP responses.