Refactor logging and improve argument handling.
Move the log-message function and the flush functionality to
src/dhcp/logging.cljs. Add a new initialization routine that takes
a configuration and stores it in the state atom (in addition to the
pending count). This simplifies later calls to log-message so the
configuration doesnt need to be passed every time.

Use the log-message in node_client for logging as well.

Change log-message to take an initial message type key that can be
either :recv, :send, :info, or :error. The :recv, :send allow
log-message to determine which direction the message is heading to print
that correctly so that it can be used for both client and server
logging. The :info and :error types are for general non-message logging.

Do better minimist arg parsing and print usage in node_client.

Rename the `disable-bcast` param/config to `disable-broadcast`.